// mocks/LOREM.mock.d.ts
export declare const LOREM: {
    button: string;
    dateFormatted: string;
    dateISO: string;
    email: () => string;
    errorMessage: (suffix?: string) => string;
    id: (suffix?: string) => string;
    label: (suffix?: string) => string;
    labelDate: (suffix?: string) => string;
    labelDescription: (suffix?: string) => string;
    labelDropdown: (suffix?: string) => string;
    labelEmail: (suffix?: string) => string;
    labelRadioButtons: (suffix?: string) => string;
    labelSlider: (suffix?: string) => string;
    link: string;
    name: (suffix?: string) => string;
    placeholder: (suffix?: string) => string;
    select: string;
    text_lg: string;
    text_md: string;
    text_xl: string;
    text_xxs: string;
    title_xl: string;
};