// test/mock_data/input_currency.d.ts
import { SelectItemShape } from "../../src/select/types";
export type MockCurrency = "USD" | "AED" | "OMR" | "SAR" | "KWD";
export type MockLocale = "en-US" | "en-GB" | "de-DE" | "fr-FR" | "ar-AE";
export declare const MOCK_CURRENCY_SELECT_ITEMS: Array<SelectItemShape<MockCurrency>>;
export declare const MOCK_LOCALE_SELECT_ITEMS: Array<SelectItemShape<MockLocale>>;